backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
requests>=2.31
supabase>=2.6
pydantic>=2.6